requirements.txt
pandas
numpy
matplotlib
gspread
oauth2client
google-auth
google-auth-oauthlib
google-auth-httplib2
prophet
openpyxl
XlsxWriter

